requirements.txt
aiofiles
deepface
face_recognition
fastapi
fpdf
fuzzywuzzy
groq
hume
joblib
langchain
langchain_community
langchain_groq
language_tool_python
matplotlib
mediapipe
nltk
numpy
openai_whisper
opencv_contrib_python
opencv_python
pandas
pdfplumber
Pillow
protobuf
pydantic
pydub
PyPDF2
python-dotenv
python_docx
pyttsx3
rapidfuzz
Requests
scikit_learn
scipy
sentence_transformers
soundfile
spacy
tenacity
textstat
torch
transformers
ultralytics
uvicorn